Alerts with swal

+++ src/app/pages/users/list/list.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
 import { User } from "src/app/models/user.model";
 import { UserService } from "src/app/services/user.service";
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: "app-list",
@@ -23,12 +25,12 @@ export class ListComponent implements OnInit {
 
   list() {
     this.service.getUsers().subscribe((data) => {
-      this.users = data.map((user: any) =>({
-          id: user?._id,
-          name: user?.name,
-          email: user?.email,
-          role: user?.role?.name.toLowerCase(),
-        }));
+      this.users = data.map((user: any) => ({
+        id: user?._id,
+        name: user?.name,
+        email: user?.email,
+        role: user?.role?.name.toLowerCase(),
+      }));
     });
   }
 
@@ -45,8 +47,26 @@ export class ListComponent implements OnInit {
   }
 
   delete(id: string) {
-    this.service.delete(id).subscribe(() => {
-      this.list();
+    Swal.fire({
+      title: '¿Estás seguro de eliminar el registro?',
+      text: "Esta acción no se puede revertir!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminar!',
+      cancelButtonText: 'No, cancelar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.service.delete(id).subscribe(() => {
+          Swal.fire(
+            'Eliminado!',
+            'El registro ha sido eliminado.',
+            'success'
+          );
+          this.ngOnInit();
+        });
+      }
     });
   }
 }
